+++ src/sdk/api/constants.ts
@@ -7,9 +7,6 @@ export const сacheSettings = {
 export const MethodTypes = {
   GET: 'GET',
   POST: 'POST',
-  PUT: 'PUT',
-  DELETE: 'DELETE',
-  PATCH: 'PATCH',
 };
 
 export const BACKEND_API_ENDPOINT = 'https://rpc.etherspot.io/data-api';
@@ -18,6 +15,7 @@ export const API_ENDPOINTS = {
   GET_ACCOUNT_BALANCES: 'account/balances',
   GET_ACCOUNT_NFTS: 'account/nfts',
   GET_TRANSACTION: 'transactions/transaction',
+  GET_TRANSACTIONS: 'transactions',
   GET_ADVANCE_ROUTES_LIFI: 'exchange/getAdvanceRoutesLiFi',
   GET_STEP_TRANSACTIONS: 'exchange/getStepTransactions',
   GET_EXCHANGE_SUPPORTED_ASSETS: 'assets/exchangeSupportedAssets',