+++ src/sdk/common/constants.ts
@@ -4,7 +4,6 @@
 export enum HeaderNames {
   AuthToken = 'x-auth-token',
   AnalyticsToken = 'x-analytics-token',
-  ProjectKey = 'x-project-key',
   ProjectMetadata = 'x-project-metadata',
 }